// src/app/shared/model/report-param.ts
export class ReportParam {
  sCompanyCode: string;
  sDate1: string;
  sDate2: string;
  sRC1: string;
  sRC2: string;
  sArticle1: string;
  sArticle2: string;
  sCustomer1: string;
  sCustomer2: string;
  sType: string;

  sMemberCodeStart: string;
  sMemberCodeEnd: string;

  sDateEnd: string;

  sRCCustomerBase1: string;
  sRCCustomerBase2: string;

  userPrint: string;
}